# shared/openapi/v1.json
{
  "components": {
    "schemas": {
      "AccountingPeriodCreate": {
        "additionalProperties": false,
        "properties": {
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "period_end": {
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date",
            "title": "Period Start",
            "type": "string"
          }
        },
        "required": [
          "period_start",
          "period_end",
          "currency"
        ],
        "title": "AccountingPeriodCreate",
        "type": "object"
      },
      "AccountingPeriodResponse": {
        "additionalProperties": false,
        "properties": {
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "period_end": {
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date",
            "title": "Period Start",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/PeriodState"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "period_start",
          "period_end",
          "currency",
          "state",
          "closed_at",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "AccountingPeriodResponse",
        "type": "object"
      },
...
      "PeriodState": {
        "enum": [
          "OPEN",
          "CLOSE_REQUESTED",
          "VALIDATING",
          "BLOCKED",
          "READY",
          "CLOSING",
          "CLOSED",
          "FAILED"
        ],
        "title": "PeriodState",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LedgerAI Backend",
    "version": "1.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/accounting-periods": {
      "get": {
        "operationId": "list_accounting_periods_api_v1_accounting_periods_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Code",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 1,
              "title": "X-Workspace-Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Legal-Entity-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Legal-Entity-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccountingPeriodResponse"
                  },
                  "title": "Response List Accounting Periods Api V1 Accounting Periods Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Accounting Periods",
        "tags": [
          "accounting-assurance",
          "audit",
          "progress"
        ]
      },
      "post": {
        "operationId": "create_accounting_period_api_v1_accounting_periods_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Workspace-Code",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 1,
              "title": "X-Workspace-Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Legal-Entity-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Legal-Entity-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountingPeriodCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountingPeriodResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Accounting Period",
        "tags": [
          "accounting-assurance",
          "audit",
          "progress"
        ]
      }
    },
    "/api/v1/accounting-periods/{period_id}/close": {
      "post": {
        "operationId": "close_period_api_v1_accounting_periods__period_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "period_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Period Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Code",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 1,
              "title": "X-Workspace-Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Legal-Entity-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Legal-Entity-Id"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloseRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Close Period",
        "tags": [
          "accounting-assurance",
          "audit",
          "progress"
        ]
      }
    },
    "/api/v1/accounting-periods/{period_id}/statements/{statement_type}": {
      "post": {
        "operationId": "capture_statement_api_v1_accounting_periods__period_id__statements__statement_type__post",
        "parameters": [
          {
            "in": "path",
            "name": "period_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Period Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "statement_type",
            "required": true,
            "schema": {
              "enum": [
                "TRIAL_BALANCE",
                "PROFIT_AND_LOSS",
                "BALANCE_SHEET"
              ],
              "title": "Statement Type",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Code",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 1,
              "title": "X-Workspace-Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Legal-Entity-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Legal-Entity-Id"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatementSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Capture Statement",
        "tags": [
          "accounting-assurance",
          "audit",
          "progress"
        ]
      }
    },
...
    "/api/v1/audit-events": {
      "get": {
        "operationId": "list_audit_events_api_v1_audit_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "resource_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Type"
            }
          },
          {
            "in": "query",
            "name": "resource_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Code",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 1,
              "title": "X-Workspace-Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Legal-Entity-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Legal-Entity-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventResponse"
                  },
                  "title": "Response List Audit Events Api V1 Audit Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Audit Events",
        "tags": [
          "accounting-assurance",
          "audit",
          "progress"
        ]
      }
    },
...
    "/api/v1/financial-statements/{snapshot_id}": {
      "get": {
        "operationId": "get_statement_api_v1_financial_statements__snapshot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Snapshot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Workspace-Code",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 1,
              "title": "X-Workspace-Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Legal-Entity-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Legal-Entity-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatementSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Statement",
        "tags": [
          "accounting-assurance",
          "audit",
          "progress"
        ]
      }
    },
...
  }
}